requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=9.0.0
moviepy>=1.0.3
psutil>=5.9.0